pyproject.toml
[project]
name = "some-asm"
version = "0.0.1"
description = "A user-friendly tool to assist in the efficient assembly and management of x86 assembly projects."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
keywords = ["assembly", "x86", "emulator", "pyqt6", "gui", "developer-tools"]
authors = [
    { name = "Yrrrrrf", email = "fer.rezac@outlook.com" },
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Assemblers",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]
dependencies = [
    "rich>=14.2.0",
    "pyqt6>=6.9.1",
    "qt-ingot>=0.0.5",
    "pydantic>=2.12.3",
    "typer>=0.20.0",
]

[project.urls]
Homepage = "https://github.com/Yrrrrrf/some-asm"
Repository = "https://github.com/Yrrrrrf/some-asm"

[tool.pyright]
venvPath = "."
venv = ".venv"
